refactor(domain): Car 관련 구현을 이동 방식에 따라 전반적으로 리팩토링
변경사항
- Car 인터페이스를 추상클래스로 변경 후 공통로직이 위치하도록 변경
- Car 추상클래스를 구현하는 클래스들은 isMovable을 구현하도록 구현
- 이에 따라 RandomMovingCar, AlwaysMovingCar, NeverMovingCar 생성
- RandomMovingCar가 GameCar의 원래 역할을 대신하도록 함
- 변경사항 관련한 테스트 구현 변경사항도 반영

Randoms의 메소드에 의해 테스트가 용이하지 않았던 도메인들의 테스트가 용이해졌다는 장점이 있다고 생각하여 리팩토링하게 되었습니다
