+++ src/main/java/racingcar/domain/AlwaysMovingCar.java
@@ -1,0 +1,12 @@
+package racingcar.domain;
+
+public class AlwaysMovingCar extends Car {
+    public AlwaysMovingCar(String name) {
+        super(name);
+    }
+
+    @Override
+    public boolean isMovable() {
+        return true;
+    }
+}
