+++ src/main/java/racingcar/domain/NeverMovingCar.java
@@ -1,0 +1,12 @@
+package racingcar.domain;
+
+public class NeverMovingCar extends Car {
+    public NeverMovingCar(String name) {
+        super(name);
+    }
+
+    @Override
+    public boolean isMovable() {
+        return false;
+    }
+}
