+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -13,29 +13,29 @@ class RacingGameTest {
 
     private static final int ZERO_ATTEMPT = 0;
     private static final int ONE_ATTEMPT = 1;
-    Car gameCar1, gameCar2;
-    List<Car> rawGameCars;
-    Cars gameCars;
+    Car alwaysMovingCar1, alwaysMovingCar2;
+    List<Car> rawCars;
+    Cars cars;
 
     @BeforeEach
     void setUp() {
-        gameCar1 = new GameCar("붕붕카");
-        gameCar2 = new GameCar("차차차");
-        rawGameCars = List.of(gameCar1, gameCar2);
-        gameCars = new GameCars(rawGameCars);
+        alwaysMovingCar1 = new AlwaysMovingCar("붕붕카");
+        alwaysMovingCar2 = new AlwaysMovingCar("차차차");
+        rawCars = List.of(alwaysMovingCar1, alwaysMovingCar2);
+        cars = new Cars(rawCars);
 
     }
 
     @Test
     void 이동_횟수_최소조건_실패_테스트() {
-        assertThatThrownBy(() -> new RacingGame(gameCars, ZERO_ATTEMPT))
+        assertThatThrownBy(() -> new RacingGame(cars, ZERO_ATTEMPT))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(IllegalArgumentExceptionType.ATTEMPT_COUNT_INIT_ERROR_MESSAGE.getMessage());
     }
 
     @Test
     void 게임_진행_성공_테스트() {
-        RacingGame racingGame = new RacingGame(gameCars, ONE_ATTEMPT);
+        RacingGame racingGame = new RacingGame(cars, ONE_ATTEMPT);
 
         SoftAssertions softAssertions = new SoftAssertions();
         softAssertions.assertThat(racingGame.isEnd()).isEqualTo(false);
@@ -46,7 +46,7 @@ class RacingGameTest {
     }
     @Test
     void 게임_진행_실패_테스트() {
-        RacingGame racingGame = new RacingGame(gameCars, ONE_ATTEMPT);
+        RacingGame racingGame = new RacingGame(cars, ONE_ATTEMPT);
         racingGame.proceed();
 
         IllegalStateException noMoreAttemptException = IllegalStateExceptionType
